已验证，可以上报数据上去，可以解析数据执行，已经解决小数点出现0.0000012的问题导致慢速运行

+++ User/app/inc/bsp_Control.h
@@ -8,7 +8,7 @@ void bsp_PidSched(void);
 void bsp_SetMotorSpeed(MotorSN sn , int32_t speed);
 int32_t bsp_MotorGetSpeed(MotorSN sn);
 int32_t bsp_MotorGetTargetSpeed(MotorSN sn);
-int32_t bsp_MotorSpeedMM2Pulse(int32_t mm);
+int32_t bsp_MotorSpeedMM2Pulse(int16_t mm);
 int32_t bsp_MotorGetPulseVector(MotorSN sn);
 
 #endif
